refactor mgmt tests
