+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_dmac_mgmt.py
@@ -11,9 +11,7 @@ from azure.core.credentials import AzureKeyCredential
 from azure.core.exceptions import ResourceNotFoundError, ClientAuthenticationError
 from azure.core.pipeline.transport import RequestsTransport
 from azure.ai.formrecognizer import (
-    FormTrainingClient,
     DocumentModelAdministrationClient,
-    FormRecognizerApiVersion,
     DocumentAnalysisApiVersion,
     ModelOperation
 )
@@ -21,7 +19,7 @@ from testcase import FormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 from preparers import FormRecognizerPreparer
 
-FormTrainingClientPreparer = functools.partial(_GlobalClientPreparer, FormTrainingClient)
+
 DocumentModelAdministrationClientPreparer = functools.partial(_GlobalClientPreparer, DocumentModelAdministrationClient)
 
 
@@ -189,102 +187,4 @@ class TestManagement(FormRecognizerTest):
                 dac.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg).wait()
                 assert dac._api_version == DocumentAnalysisApiVersion.V2021_09_30_PREVIEW
             dtc.get_account_info()
-            assert transport.session is not None
-
-    # --------------------------------------- BACK COMPATABILITY TESTS ---------------------------------------
-
-    @FormRecognizerPreparer()
-    @FormTrainingClientPreparer(client_kwargs={"api_version": "2.1"})
-    def test_account_properties_v2(self, client):
-        properties = client.get_account_properties()
-
-        assert properties.custom_model_limit
-        assert properties.custom_model_count
-
-    @pytest.mark.skip("service is returning null for some models")
-    @FormRecognizerPreparer()
-    @FormTrainingClientPreparer(client_kwargs={"api_version": "2.1"})
-    def test_mgmt_model_labeled_v2(self, client, formrecognizer_storage_container_sas_url_v2):
-
-        poller = client.begin_training(formrecognizer_storage_container_sas_url_v2, use_training_labels=True)
-        labeled_model_from_train = poller.result()
-
-        labeled_model_from_get = client.get_custom_model(labeled_model_from_train.model_id)
-
-        assert labeled_model_from_train.model_id == labeled_model_from_get.model_id
-        assert labeled_model_from_train.status == labeled_model_from_get.status
-        assert labeled_model_from_train.training_started_on == labeled_model_from_get.training_started_on
-        assert labeled_model_from_train.training_completed_on == labeled_model_from_get.training_completed_on
-        assert labeled_model_from_train.errors == labeled_model_from_get.errors
-        for a, b in zip(labeled_model_from_train.training_documents, labeled_model_from_get.training_documents):
-            assert a.name == b.name
-            assert a.errors == b.errors
-            assert a.page_count == b.page_count
-            assert a.status == b.status
-        for a, b in zip(labeled_model_from_train.submodels, labeled_model_from_get.submodels):
-            for field1, field2 in zip(a.fields.items(), b.fields.items()):
-                assert a.fields[field1[0]].name == b.fields[field2[0]].name
-                assert a.fields[field1[0]].accuracy == b.fields[field2[0]].accuracy
-
-        models_list = client.list_custom_models()
-        for model in models_list:
-            assert model.model_id
-            assert model.status
-            assert model.training_started_on
-            assert model.training_completed_on
-
-        client.delete_model(labeled_model_from_train.model_id)
-
-        with pytest.raises(ResourceNotFoundError):
-            client.get_custom_model(labeled_model_from_train.model_id)
-
-    @pytest.mark.skip("service is returning null for some models")
-    @FormRecognizerPreparer()
-    @FormTrainingClientPreparer(client_kwargs={"api_version": "2.1"})
-    def test_mgmt_model_unlabeled_v2(self, client, formrecognizer_storage_container_sas_url_v2):
-
-        poller = client.begin_training(formrecognizer_storage_container_sas_url_v2, use_training_labels=False)
-        unlabeled_model_from_train = poller.result()
-
-        unlabeled_model_from_get = client.get_custom_model(unlabeled_model_from_train.model_id)
-
-        assert unlabeled_model_from_train.model_id == unlabeled_model_from_get.model_id
-        assert unlabeled_model_from_train.status == unlabeled_model_from_get.status
-        assert unlabeled_model_from_train.training_started_on == unlabeled_model_from_get.training_started_on
-        assert unlabeled_model_from_train.training_completed_on == unlabeled_model_from_get.training_completed_on
-        assert unlabeled_model_from_train.errors == unlabeled_model_from_get.errors
-        for a, b in zip(unlabeled_model_from_train.training_documents, unlabeled_model_from_get.training_documents):
-            assert a.name == b.name
-            assert a.errors == b.errors
-            assert a.page_count == b.page_count
-            assert a.status == b.status
-        for a, b in zip(unlabeled_model_from_train.submodels, unlabeled_model_from_get.submodels):
-            for field1, field2 in zip(a.fields.items(), b.fields.items()):
-                assert a.fields[field1[0]].label == b.fields[field2[0]].label
-
-        models_list = client.list_custom_models()
-        for model in models_list:
-            assert model.model_id
-            assert model.status
-            assert model.training_started_on
-            assert model.training_completed_on
-
-        client.delete_model(unlabeled_model_from_train.model_id)
-
-        with pytest.raises(ResourceNotFoundError):
-            client.get_custom_model(unlabeled_model_from_train.model_id)
-
-    @FormRecognizerPreparer()
-    def test_get_form_recognizer_client_v2(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        transport = RequestsTransport()
-        ftc = FormTrainingClient(endpoint=formrecognizer_test_endpoint, credential=AzureKeyCredential(formrecognizer_test_api_key), transport=transport, api_version="2.1")
-
-        with ftc:
-            ftc.get_account_properties()
-            assert transport.session is not None
-            with ftc.get_form_recognizer_client() as frc:
-                assert transport.session is not None
-                frc.begin_recognize_receipts_from_url(self.receipt_url_jpg).wait()
-                assert frc._api_version == FormRecognizerApiVersion.V2_1
-            ftc.get_account_properties()
             assert transport.session is not None
